+++ src/Entity/DailyChange.php
@@ -2,6 +2,7 @@
 
 namespace App\Entity;
 
+use App\Entity\Traits\ChangeTrait;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -10,9 +11,12 @@ use Doctrine\ORM\Mapping as ORM;
  */
 class DailyChange
 {
+    use ChangeTrait;
+
     /**
      * @ORM\Id()
      * @ORM\GeneratedValue()
+     * @var int
      * @ORM\Column(type="integer")
      */
     private $id;
@@ -26,46 +30,6 @@ class DailyChange
     private $dailyStat;
 
     /**
-     * @ORM\Column(type="integer")
-     */
-    private $new;
-
-    /**
-     * @ORM\Column(type="float")
-     */
-    private $newPercent;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    private $deaths;
-
-    /**
-     * @ORM\Column(type="float")
-     */
-    private $deathsPercent;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    private $recovered;
-
-    /**
-     * @ORM\Column(type="float")
-     */
-    private $recoveredPercent;
-
-    /**
-     * @ORM\Column(type="integer")
-     */
-    private $serious;
-
-    /**
-     * @ORM\Column(type="float")
-     */
-    private $seriousPercent;
-
-    /**
      * DailyChange constructor.
      *
      * @param $dailyStat
@@ -76,15 +40,6 @@ class DailyChange
             $dailyStat->setDailyChange($this);
             $this->dailyStat = $dailyStat;
         }
-
-        $this->new = 0;
-        $this->newPercent = 0;
-        $this->deaths = 0;
-        $this->deathsPercent = 0;
-        $this->recovered = 0;
-        $this->recoveredPercent = 0;
-        $this->serious = 0;
-        $this->seriousPercent = 0;
     }
 
 
@@ -94,226 +49,6 @@ class DailyChange
     public function getId()
     {
         return $this->id;
-    }
-
-    /**
-     * @return int
-     */
-    public function getNew(): int
-    {
-        return $this->new;
-    }
-
-    /**
-     * @param mixed $new
-     */
-    public function setNew($new): void
-    {
-        $this->new = $new;
-    }
-
-    /**
-     * @param integer $new
-     * @param integer $percent
-     */
-    public function setNewData($new, $percent): void
-    {
-        $this->new = $new;
-        $this->newPercent = $percent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getNewPercent(): int
-    {
-        return $this->newPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getNewValue(): ?string
-    {
-        $value = abs($this->new);
-        if ($this->newPercent !== 0) {
-            $value = abs($this->newPercent)."%";
-        }
-
-        return $value;
-    }
-
-    /**
-     * @param mixed $newPercent
-     */
-    public function setNewPercent($newPercent): void
-    {
-        $this->newPercent = $newPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getDeaths(): int
-    {
-        return $this->deaths;
-    }
-
-    /**
-     * @param mixed $deaths
-     */
-    public function setDeaths($deaths): void
-    {
-        $this->deaths = $deaths;
-    }
-
-    /**
-     * @param integer $deaths
-     * @param integer $percent
-     */
-    public function setDeathsData($deaths, $percent): void
-    {
-        $this->deaths = $deaths;
-        $this->deathsPercent = $percent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getDeathsPercent(): int
-    {
-        return $this->deathsPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getDeathsValue(): ?string
-    {
-        $value = abs($this->deaths);
-        if ($this->deathsPercent !== 0) {
-            $value = abs($this->deathsPercent)."%";
-        }
-
-        return $value;
-    }
-
-    /**
-     * @param mixed $deathsPercent
-     */
-    public function setDeathsPercent($deathsPercent): void
-    {
-        $this->deathsPercent = $deathsPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getRecovered(): int
-    {
-        return $this->recovered;
-    }
-
-    /**
-     * @param mixed $recovered
-     */
-    public function setRecovered($recovered): void
-    {
-        $this->recovered = $recovered;
-    }
-
-    /**
-     * @param integer $recovered
-     * @param integer $percent
-     */
-    public function setRecoveredData($recovered, $percent): void
-    {
-        $this->recovered = $recovered;
-        $this->recoveredPercent = $percent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getRecoveredPercent(): int
-    {
-        return $this->recoveredPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getRecoveredValue(): ?string
-    {
-        $value = abs($this->recovered);
-        if ($this->recoveredPercent !== 0) {
-            $value = abs($this->recoveredPercent)."%";
-        }
-
-        return $value;
-    }
-
-    /**
-     * @param mixed $recoveredPercent
-     */
-    public function setRecoveredPercent($recoveredPercent): void
-    {
-        $this->recoveredPercent = $recoveredPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getSerious(): int
-    {
-        return $this->serious;
-    }
-
-    /**
-     * @param mixed $serious
-     */
-    public function setSerious($serious): void
-    {
-        $this->serious = $serious;
-    }
-
-    /**
-     * @param integer $serious
-     * @param integer $percent
-     */
-    public function setSeriousData($serious, $percent): void
-    {
-        $this->serious = $serious;
-        $this->seriousPercent = $percent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getSeriousPercent(): int
-    {
-        return $this->seriousPercent;
-    }
-
-    /**
-     * @return int
-     */
-    public function getSeriousValue(): ?string
-    {
-        $value = abs($this->serious);
-        if ($this->seriousPercent !== 0) {
-            $value = abs($this->seriousPercent)."%";
-        }
-
-        return $value;
-    }
-
-    /**
-     * @param mixed $seriousPercent
-     */
-    public function setSeriousPercent($seriousPercent): void
-    {
-        $this->seriousPercent = $seriousPercent;
     }
 
     public function getDailyStat(): ?DailyStat
