+++ src/Twig/NumberExtension.php
@@ -11,6 +11,7 @@ class NumberExtension extends AbstractExtension
     {
         return [
             new TwigFilter('number', [$this, 'formatNumber']),
+            new TwigFilter('number_signed', [$this, 'signedNumber']),
         ];
     }
 
@@ -19,4 +20,16 @@ class NumberExtension extends AbstractExtension
     {
         return number_format($value, 0, ".", " ");
     }
+
+
+    public function signedNumber($value)
+    {
+        if (0 === $value) {
+            return 0;
+        }
+
+        return ($value > 0 ? '+' : '').$this->formatNumber($value);
+    }
+
+
 }
